создает и подключает модули
